+++ src/main/java/com/hongik/graduationproject/domain/auth/oauth/KaKaoProfile.java
@@ -1,4 +1,4 @@
-package com.hongik.graduationproject.domain.oauth;
+package com.hongik.graduationproject.domain.auth.oauth;
 
 import lombok.Data;
 
@@ -30,6 +30,7 @@ public class KaKaoProfile {
 
         @Data
         public class Profile {
+
             public String nickname;
             public String thumbnail_image_url;
             public String profile_image_url;
