+++ src/main/java/com/hongik/graduationproject/domain/auth/oauth/UserId.java
@@ -1,4 +1,4 @@
-package com.hongik.graduationproject.domain.oauth;
+package com.hongik.graduationproject.domain.auth.oauth;
 
 import lombok.Value;
