+++ src/main/java/com/hongik/graduationproject/domain/dto/KaKaoRequestDto.java
@@ -1,4 +1,4 @@
-package com.hongik.graduationproject.domain.oauth;
+package com.hongik.graduationproject.domain.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class KaKaoRequestDto {
+public class KaKaoRequestDto extends AuthRequestDto {
 
     String accessToken;
     String refreshToken;
